Modify the v13 migration file to not delete the admin account in production. Update the UserSeederRunner for development.

+++ src/main/resources/db/migration/V13__un_seed_prev_data.sql
@@ -4,7 +4,9 @@ DELETE FROM appointments;
 DELETE FROM users_services;
 DELETE FROM services;
 DELETE FROM work_spaces;
-DELETE FROM staff_details;
-DELETE FROM users;
+DELETE FROM staff_details WHERE user_id != (select id from users where email = 'admin@abv.bg');
+DELETE FROM users WHERE email != 'admin@abv.bg';
+
+UPDATE users SET email = 'admin@flow.com' WHERE email = 'admin@abv.bg';
 
 COMMIT;
